% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-graphics.R
\name{include_graphics2}
\alias{include_graphics2}
\title{A wrapper of knitr::include_graphics with more automatic file conversions}
\usage{
include_graphics2(path, auto_pdf = TRUE, auto_png = TRUE, dpi = NULL)
}
\arguments{
\item{path}{A character vector of image paths.}

\item{auto_pdf}{Convert svg to pdf when knitting pdf (default: `TRUE`).}

\item{auto_png}{Convert svg to png when knitting docx (default: `TRUE`)}

\item{dpi}{DPI (dots per inch) value. Used to calculate the output
width (in inches) of the images. This will be their actual width in pixels,
divided by \code{dpi}. If not provided, the chunk option \code{dpi} is used; if
\code{NA}, the output width will not be calculated.}
}
\value{
The same as the input character vector \code{path} but it is marked
  with special internal S3 classes so that \pkg{knitr} will convert the file
  paths to proper output code according to the output format.
}
\description{
When plots are not generated from R code, there is no way for \pkg{knitr} to
capture plots automatically. In this case, you may generate the images
manually and pass their file paths to this function to include them in the
output. The major advantage of using this function is that it is portable in
the sense that it works for all document formats that \pkg{knitr} supports,
so you do not need to think if you have to use, for example, LaTeX or
Markdown syntax, to embed an external image. Chunk options related to
graphics output that work for normal R plots also work for these images, such
as \code{out.width} and \code{out.height}.
}
