% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engines.R
\name{knit-engines}
\alias{knit-engines}
\alias{eng_text}
\alias{eng_glue}
\alias{eng_opts}
\title{Knit engines}
\usage{
eng_text(options)

eng_glue(options)

eng_opts(options)
}
\description{
Additional engines for Rmd chunks
}
\section{}{
NA
}

