% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-hooks.R
\name{knit-hooks}
\alias{knit-hooks}
\alias{hook_time}
\alias{hook_src}
\title{Knit hooks}
\usage{
hook_time(before, options, envir)

hook_src(default = knit_hooks$get("source"))
}
\arguments{
\item{before, options, envir}{Required parameters for chunk hooks.}

\item{default}{A default function to render source chunk.}
}
\description{
Knit hooks
}
\section{time}{

  Specify `TRUE` to measure elapsed time by evaluating a chunk.
}

\section{opts.include}{

  Specify `TRUE` to output source chunks together with their chunk options.
}

