% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engines.R
\name{opts_src}
\alias{opts_src}
\title{Returns chunk options of a source chunk}
\usage{
opts_src(options)
}
\arguments{
\item{options}{Chunk options}
}
\description{
Returns chunk options of a source chunk
}
